backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
greenlet>=3.0.0
supabase==2.15.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
slowapi==0.1.9
google-cloud-logging==3.8.0
structlog==23.2.0
prometheus-client==0.19.0
pytest>=8.0.0
httpx>=0.26.0,<0.29.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
numpy>=1.22.0
torch>=2.0.0
transformers>=4.52.0
google-cloud-documentai>=3.5.0
google-generativeai>=0.4.2
google-cloud-storage==2.13.0
google-cloud-tasks==2.13.1
google-cloud-secret-manager==2.16.4
Pillow==10.1.0
reportlab>=4.4.1
pandas>=1.4.0
requests==2.31.0
aiofiles==23.2.1
pytz==2023.3
python-dateutil==2.8.2
pgvector==0.2.4
bcrypt==4.1.1
pydantic[email]==2.5.0
aiohttp>=3.11.18,<4.0.0
spacy>=3.7.0
